stock-portfolio-project/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0

numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
statsmodels==0.14.6
yfinance==1.2.0
requests==2.32.5

# Required for the RNN forecasting endpoint.
tensorflow==2.20.0
